% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_expr_template.R
\name{bf_expr_template}
\alias{bf_expr_template}
\title{Expression template for Bayes Factor results}
\usage{
bf_expr_template(
  top.text,
  bf.value,
  bf.prior = 0.707,
  prior.type = quote(italic("r")["Cauchy"]^"JZS"),
  estimate.type = quote(delta),
  estimate,
  estimate.LB,
  estimate.UB,
  centrality = "median",
  conf.level = 0.95,
  conf.method = "HDI",
  k = 2L,
  ...
)
}
\arguments{
\item{top.text}{Text to display as top.text (will be displayed on top of the
Bayes Factor top.text/message).}

\item{bf.value}{Numeric value corresponding to Bayes Factor.}

\item{bf.prior}{A number between \code{0.5} and \code{2} (default \code{0.707}), the prior
width to use in calculating Bayes factors.}

\item{prior.type}{A character that specifies the prior type.}

\item{estimate.type}{A character that specifies the relevant effect size.}

\item{estimate, estimate.LB, estimate.UB}{Values of posterior estimates and
their credible intervals.}

\item{centrality}{The point-estimates (centrality indices) to compute.
Character (vector) or list with one or more of these options: \code{"median"},
\code{"mean"}, \code{"MAP"} or \code{"all"}.}

\item{conf.level}{Confidence/Credible Interval (CI) level. Default to \code{0.95}
(\verb{95\%}).}

\item{conf.method}{The type of index used for Credible Interval. Can be
\code{"hdi"} (default, see \code{\link[bayestestR:hdi]{bayestestR::hdi()}}), \code{"eti"} (see
\code{\link[bayestestR:eti]{bayestestR::eti()}}) or \code{"si"} (see \code{\link[bayestestR:si]{bayestestR::si()}}).}

\item{k}{Number of digits after decimal point (should be an integer)
(Default: \code{k = 2L}).}

\item{...}{Currently ignored.}
}
\description{
Expression template for Bayes Factor results
}
