% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bf_tests.R
\name{meta_data_check}
\alias{meta_data_check}
\title{Helper function to check column names for meta-analysis.}
\usage{
meta_data_check(data)
}
\arguments{
\item{data}{A dataframe. It \strong{must} contain columns named \code{estimate} (effect
sizes or outcomes)  and \code{std.error} (corresponding standard errors). These
two columns will be used for \code{y}  and \code{SE} arguments in
\code{metaBMA::meta_random}.}
}
\description{
Helper function to check column names for meta-analysis.
}
