% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_bf_tests.R
\name{bf_expr}
\alias{bf_expr}
\title{Prepare expression for Bayes Factor results}
\usage{
bf_expr(
  bf.object,
  k = 2L,
  conf.level = 0.95,
  conf.method = "hdi",
  centrality = "median",
  top.text = NULL,
  anova.design = FALSE,
  ...
)
}
\arguments{
\item{bf.object}{An object from \code{BayesFactor} package.}

\item{k}{Number of digits after decimal point (should be an integer)
(Default: \code{k = 2L}).}

\item{conf.level}{Confidence/Credible Interval (CI) level. Default to \code{0.95}
(\verb{95\%}).}

\item{conf.method}{The type of index used for Credible Interval. Can be
\code{"hdi"} (default, see \code{\link[bayestestR:hdi]{bayestestR::hdi()}}), \code{"eti"} (see
\code{\link[bayestestR:eti]{bayestestR::eti()}}) or \code{"si"} (see \code{\link[bayestestR:si]{bayestestR::si()}}).}

\item{centrality}{The point-estimates (centrality indices) to compute.
Character (vector) or list with one or more of these options: \code{"median"},
\code{"mean"}, \code{"MAP"} or \code{"all"}.}

\item{top.text}{Text to display as top.text (will be displayed on top of the
Bayes Factor top.text/message).}

\item{anova.design}{Whether the object is from \code{BayesFactor::anovaBF}
(default: \code{FALSE}). The expression is different for anova designs because
not all details are available.}

\item{...}{Additional arguments passed to
\code{\link[parameters:model_parameters.BFBayesFactor]{parameters::model_parameters.BFBayesFactor()}}.}
}
\description{
Convenience function to create an expression with Bayes
Factor results.
}
\examples{
# for reproducibility
set.seed(123)
library(tidyBF)

# creating expression
bf_expr(
  bf.object = BayesFactor::correlationBF(
    x = iris$Sepal.Length,
    y = iris$Petal.Length
  ),
  k = 3,
  top.text = "Note: Iris dataset"
)
}
