% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_oneway_anova.R
\name{bf_oneway_anova}
\alias{bf_oneway_anova}
\title{Bayesian one-way analysis of variance (ANOVA)}
\usage{
bf_oneway_anova(
  data,
  x,
  y,
  bf.prior = 0.707,
  top.text = NULL,
  output = "dataframe",
  paired = FALSE,
  k = 2L,
  ...
)
}
\arguments{
\item{data}{A dataframe (or a tibble) from which variables specified are to
be taken. A matrix or tables will \strong{not} be accepted.}

\item{x}{The grouping variable from the dataframe \code{data}.}

\item{y}{The response (a.k.a. outcome or dependent) variable from the
dataframe \code{data}.}

\item{bf.prior}{A number between \code{0.5} and \code{2} (default \code{0.707}), the prior
width to use in calculating Bayes factors.}

\item{top.text}{Text to display as top.text (will be displayed on top of the
Bayes Factor top.text/message).}

\item{output}{If \code{"expression"}, will return expression with statistical
details, while \code{"dataframe"} (default) will return a dataframe containing the
results.}

\item{paired}{Decides whether the design is repeated measures or not
(Default: \code{FALSE}).}

\item{k}{Number of digits after decimal point (should be an integer)
(Default: \code{k = 2L}).}

\item{...}{Additional arguments passed to
\code{\link[parameters:model_parameters.BFBayesFactor]{parameters::model_parameters.BFBayesFactor()}}.}
}
\description{
Bayesian one-way analysis of variance (ANOVA)
}
\examples{
# setup
set.seed(123)

# between-subjects -------------------------------------------------------

# to get dataframe
bf_oneway_anova(
  data = iris,
  x = Species,
  y = Sepal.Length,
  bf.prior = 0.8,
  output = "dataframe"
)

# within-subjects -------------------------------------------------------

# to get expression
bf_oneway_anova(
  data = bugs_long,
  x = condition,
  y = desire,
  paired = TRUE,
  output = "expression"
)
}
\seealso{
\code{\link{bf_contingency_tab}}, \code{\link{bf_corr_test}},
\code{\link{bf_ttest}}
}
